refactor: move `Modifier` from `Ast` to `shared`
