+++ main/src/ca/uwaterloo/flix/language/ast/shared/Modifier.scala
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2024 Holger Dal Mogensen
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ca.uwaterloo.flix.language.ast.shared
+
+/**
+  * A common super-type for modifiers.
+  */
+sealed trait Modifier
+
+object Modifier {
+
+  /**
+    * The lawful modifier.
+    */
+  case object Lawful extends Modifier
+
+  /**
+    * The mutable modifier.
+    */
+
+  case object Mutable extends Modifier
+
+  /**
+    * The override modifier.
+    */
+  case object Override extends Modifier
+
+  /**
+    * The public modifier.
+    */
+  case object Public extends Modifier
+
+  /**
+    * The sealed modifier.
+    */
+  case object Sealed extends Modifier
+
+  /**
+    * The synthetic modifier.
+    */
+  case object Synthetic extends Modifier
+
+}
