+++ main/src/ca/uwaterloo/flix/language/phase/Weeder2.scala
@@ -19,7 +19,7 @@ import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.language.CompilationMessage
 import ca.uwaterloo.flix.language.ast.Ast.*
 import ca.uwaterloo.flix.language.ast.SyntaxTree.{Tree, TreeKind}
-import ca.uwaterloo.flix.language.ast.shared.{Annotation, Annotations, CheckedCastType, Constant, Denotation, Doc, Fixity, Modifiers, Polarity}
+import ca.uwaterloo.flix.language.ast.shared.{Annotation, Annotations, CheckedCastType, Constant, Denotation, Doc, Fixity, Modifier, Modifiers, Polarity}
 import ca.uwaterloo.flix.language.ast.{Ast, ChangeSet, Name, ReadAst, SemanticOp, SourceLocation, Symbol, SyntaxTree, Token, TokenKind, WeededAst}
 import ca.uwaterloo.flix.language.dbg.AstPrinter.*
 import ca.uwaterloo.flix.language.errors.ParseError.*
@@ -740,14 +740,14 @@ object Weeder2 {
       }
     }
 
-    private def visitModifier(token: Token, allowed: Set[TokenKind]): Validation[Ast.Modifier, CompilationMessage] = {
+    private def visitModifier(token: Token, allowed: Set[TokenKind]): Validation[Modifier, CompilationMessage] = {
       val mod = token.kind match {
-        // TODO: there is no Ast.Modifier for 'inline'
-        case TokenKind.KeywordSealed => Validation.success(Ast.Modifier.Sealed)
-        case TokenKind.KeywordLawful => Validation.success(Ast.Modifier.Lawful)
-        case TokenKind.KeywordPub => Validation.success(Ast.Modifier.Public)
-        case TokenKind.KeywordMut => Validation.success(Ast.Modifier.Mutable)
-        case TokenKind.KeywordOverride => Validation.success(Ast.Modifier.Override)
+        // TODO: there is no Modifier for 'inline'
+        case TokenKind.KeywordSealed => Validation.success(Modifier.Sealed)
+        case TokenKind.KeywordLawful => Validation.success(Modifier.Lawful)
+        case TokenKind.KeywordPub => Validation.success(Modifier.Public)
+        case TokenKind.KeywordMut => Validation.success(Modifier.Mutable)
+        case TokenKind.KeywordOverride => Validation.success(Modifier.Override)
         case kind => throw InternalCompilerException(s"Parser passed unknown modifier '$kind'", token.mkSourceLocation())
       }
       if (!allowed.contains(token.kind)) {
